+++ src/main/java/com/project/growfit/domain/auth/controller/AuthChildController.java
@@ -1,10 +1,10 @@
-package com.project.growfit.domain.User.controller;
+package com.project.growfit.domain.auth.controller;
 
-import com.project.growfit.domain.User.dto.request.AuthChildRequestDto;
-import com.project.growfit.domain.User.dto.request.FindChildPasswordRequestDto;
+import com.project.growfit.domain.auth.dto.request.AuthChildRequestDto;
+import com.project.growfit.domain.auth.dto.request.FindChildPasswordRequestDto;
 import com.project.growfit.domain.User.dto.response.ChildInfoResponseDto;
-import com.project.growfit.domain.User.dto.response.ChildResponseDto;
-import com.project.growfit.domain.User.service.AuthChildService;
+import com.project.growfit.domain.auth.dto.response.ChildResponseDto;
+import com.project.growfit.domain.auth.service.AuthChildService;
 import com.project.growfit.global.response.ResultCode;
 import com.project.growfit.global.response.ResultResponse;
 import io.swagger.v3.oas.annotations.Operation;
