+++ src/main/java/com/project/growfit/domain/auth/controller/OAuthController.java
@@ -1,7 +1,7 @@
-package com.project.growfit.domain.User.controller;
+package com.project.growfit.domain.auth.controller;
 
-import com.project.growfit.domain.User.dto.response.ParentLoginResponseDto;
-import com.project.growfit.domain.User.service.OauthService;
+import com.project.growfit.domain.auth.dto.response.ParentLoginResponseDto;
+import com.project.growfit.domain.auth.service.OauthService;
 import com.project.growfit.global.exception.BusinessException;
 import com.project.growfit.global.response.ResultCode;
 import com.project.growfit.global.response.ResultResponse;
