+++ src/main/java/com/project/growfit/domain/auth/dto/request/AuthChildRequestDto.java
@@ -1,4 +1,4 @@
-package com.project.growfit.domain.User.dto.request;
+package com.project.growfit.domain.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
