+++ src/main/java/com/project/growfit/domain/auth/service/AuthChildService.java
@@ -1,9 +1,9 @@
-package com.project.growfit.domain.User.service;
+package com.project.growfit.domain.auth.service;
 
-import com.project.growfit.domain.User.dto.request.AuthChildRequestDto;
-import com.project.growfit.domain.User.dto.request.FindChildPasswordRequestDto;
+import com.project.growfit.domain.auth.dto.request.AuthChildRequestDto;
+import com.project.growfit.domain.auth.dto.request.FindChildPasswordRequestDto;
 import com.project.growfit.domain.User.dto.response.ChildInfoResponseDto;
-import com.project.growfit.domain.User.dto.response.ChildResponseDto;
+import com.project.growfit.domain.auth.dto.response.ChildResponseDto;
 import jakarta.servlet.http.HttpServletResponse;
 
 public interface AuthChildService {
