+++ src/main/java/com/project/growfit/domain/auth/service/AuthParentService.java
@@ -1,9 +1,9 @@
-package com.project.growfit.domain.User.service;
+package com.project.growfit.domain.auth.service;
 
 import com.google.zxing.WriterException;
-import com.project.growfit.domain.User.dto.request.AuthParentRequestDto;
+import com.project.growfit.domain.auth.dto.request.AuthParentRequestDto;
 import com.project.growfit.domain.User.dto.response.ChildInfoResponseDto;
-import com.project.growfit.domain.User.dto.response.ChildQrCodeResponseDto;
+import com.project.growfit.domain.auth.dto.response.ChildQrCodeResponseDto;
 
 public interface AuthParentService {
     ChildInfoResponseDto registerChild(AuthParentRequestDto request);
