+++ src/main/java/com/project/growfit/domain/auth/service/impl/AuthChildServiceImpl.java
@@ -1,12 +1,12 @@
-package com.project.growfit.domain.User.service.impl;
+package com.project.growfit.domain.auth.service.impl;
 
-import com.project.growfit.domain.User.dto.request.AuthChildRequestDto;
-import com.project.growfit.domain.User.dto.request.FindChildPasswordRequestDto;
+import com.project.growfit.domain.auth.dto.request.AuthChildRequestDto;
+import com.project.growfit.domain.auth.dto.request.FindChildPasswordRequestDto;
 import com.project.growfit.domain.User.dto.response.ChildInfoResponseDto;
-import com.project.growfit.domain.User.dto.response.ChildResponseDto;
+import com.project.growfit.domain.auth.dto.response.ChildResponseDto;
 import com.project.growfit.domain.User.entity.Child;
 import com.project.growfit.domain.User.repository.ChildRepository;
-import com.project.growfit.domain.User.service.AuthChildService;
+import com.project.growfit.domain.auth.service.AuthChildService;
 import com.project.growfit.global.auth.cookie.CookieService;
 import com.project.growfit.global.auth.jwt.JwtProvider;
 import com.project.growfit.global.auth.service.AuthenticatedUserProvider;
